refactor(70068-models): refactored models into distinct files
