+++ scheduling_resource_allocation/common/common.py
@@ -1,4 +1,4 @@
-from common.models import Node
+from models.Node import Node
 
 def tardiness_cost_fn(C_j: int, node: Node):
     return max(0, C_j - node.due_date)
