+++ scheduling_resource_allocation/models/Node.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+from models.OperationType import OperationType
+
+
+class Node(BaseModel):
+    id: int
+    operation: OperationType
+    due_date: int
