+++ scheduling_resource_allocation/models/OperationType.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class OperationType(Enum):
+    ONNX = "onnx"
+    MUSE = "muse"
+    EMBOSS = "emboss"
+    BLUR = "blur"
+    WAVE = "wave"
+    VII = "vii"
+    NIGHT = "night"
