+++ scheduling_resource_allocation/models/TardySchedule.py
@@ -1,0 +1,8 @@
+from common.common import tardiness_cost_fn
+from models.Schedule import Schedule
+from models.DirectedAcyclicGraph import DirectedAcyclicGraph
+
+
+class TardySchedule(Schedule):
+    def __init__(self, schedule: list[int], dag: DirectedAcyclicGraph):
+        super().__init__(schedule, dag, tardiness_cost_fn)
